Added Key Security, API querying with basic modeling, got data flowing properly
Added a dependency called cocoapods keys which obfiscates the client id and keys so that I don't commit my api secrets to SC, and created the basic API querying and JSON parsing with the models required for the incoming data. At this point, the data seems to be flowing properly.